+++ vms/saevm/cchain/vm.go
@@ -11,6 +11,7 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"net/http"
 	"slices"
 	"sync"
 	"time"
@@ -24,6 +25,7 @@ import (
 	"github.com/ava-labs/avalanchego/api"
 	"github.com/ava-labs/avalanchego/database/prefixdb"
 	"github.com/ava-labs/avalanchego/graft/coreth/plugin/evm/customtypes"
+	"github.com/ava-labs/avalanchego/graft/evm/utils/rpc"
 	"github.com/ava-labs/avalanchego/network/p2p"
 	"github.com/ava-labs/avalanchego/network/p2p/gossip"
 	"github.com/ava-labs/avalanchego/snow"
@@ -367,6 +369,42 @@ func (vm *VM) ParseBlock(ctx context.Context, buf []byte) (*blocks.Block, error)
 		return nil, fmt.Errorf("%w: have %x, want %x", errExtDataHashMismatch, got, wantHeaderHash)
 	}
 	return b, nil
+}
+
+const (
+	avaxServiceName       = "avax"
+	avaxHTTPExtensionPath = "/" + avaxServiceName
+)
+
+var handlerPaths = append(slices.Clone(sae.HandlerPaths), avaxHTTPExtensionPath)
+
+// CreateHandlers returns the HTTP handlers exposed by the underlying SAE VM
+// augmented with the avax service. None of the handlers are usable until after
+// the [VM] is set as bootstrapping/normal operation.
+func (vm *VM) CreateHandlers(context.Context) (map[string]http.Handler, error) {
+	return vm.handlers.AsInterface(), nil
+}
+
+// setHandlers initializes the lazy handlers with the real implementations.
+func (vm *VM) setHandlers(ctx context.Context) error {
+	m, err := vm.VM.CreateHandlers(ctx)
+	if err != nil {
+		return fmt.Errorf("creating SAE handlers: %w", err)
+	}
+
+	service, err := newService(vm.ctx, vm.gossipSet, vm.pushGossiper, vm.state)
+	if err != nil {
+		return fmt.Errorf("creating avax service: %w", err)
+	}
+	handler, err := rpc.NewHandler(avaxServiceName, service)
+	if err != nil {
+		return fmt.Errorf("creating avax RPC handler: %w", err)
+	}
+
+	m[avaxHTTPExtensionPath] = handler
+
+	vm.handlers.Set(m)
+	return nil
 }
 
 // Shutdown releases every resource allocated by the [VM] in reverse order.
